Rename jsonschema to specschema
