+++ code/go/internal/specschema/testdata/simple-spec/docs/spec.yml
