+++ code/go/internal/specschema/testdata/simple-spec/manifest.spec.yml
