+++ code/go/internal/specschema/testdata/simple-spec/spec.yml
